cleaned up folder and renamed root
