+++ index.html
@@ -8,7 +8,7 @@
     <link href="https://fonts.googleapis.com/css2?family=Ubuntu:ital,wght@0,300;0,400;0,500;0,700;1,300;1,400;1,500;1,700&display=swap" rel="stylesheet">
 
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Helvar Service Tickets</title>
+    <title>Helvar Support Portal</title>
   </head>
   <body>
     <div id="app"></div>
