+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "login-page",
+  "name": "helvar-service-tickets",
   "version": "0.0.0",
   "private": true,
   "type": "module",
