+++ src/views/auth-views/AccountCreationConfirmationView.vue
@@ -11,7 +11,7 @@
         <div id="auth-header-text">
             <div id="auth-header-title">
                 <LogoIcon />
-                <span style="font-weight: 800; font-size: 20px">HELVAR SERVICE TICKETS</span>
+                <span style="font-weight: 800; font-size: 20px">HELVAR SUPPORT PORTAL</span>
             </div>
 
             <span id="auth-header-subtext">
