+++ src/views/auth-views/LoginView.vue
@@ -1,37 +1,25 @@
 <script setup>
-    import { ref } from 'vue';
-    import RegisterForm from '@/components/user-auth/RegisterForm.vue';
+    import LoginForm from '@/components/user-auth/LoginForm.vue';
     import LogoIcon from '@/components/graphic-items/LogoIcon.vue';
+
     import { useI18n } from 'vue-i18n';
-
     const { t } = useI18n();
-
-    const formProgress = ref(0);
-    const emit = defineEmits(['form-progress']);
-
-    function handleFormProgress(value) {
-        formProgress.value = value;
-        emit('form-progress', value);
-    }
 </script>
 
 <template>
-    <div id="register-page-content">
-        <div 
-            id="auth-header-text"
-            style="align-self: flex-start"
-        >
+    <div id="login-page-content">
+        <div id="auth-header-text">
             <div id="auth-header-title">
                 <LogoIcon />
-                <span :style="{ 'font-weight' : 800, 'font-size' : '20px' }">HELVAR SERVICE TICKETS</span>
+                <span style="font-weight: 800; font-size: 20px">HELVAR SUPPORT PORTAL</span>
             </div>
 
             <span id="auth-header-subtext">
-                {{ t('registerHeaderSubtext') }}
+                {{ t('loginHeaderSubtext') }}
             </span>
-        </div> 
+        </div>
 
-        <RegisterForm @form-progress="handleFormProgress" />
+        <LoginForm />
 
         <div id="horizontal-divider">
             <span class="horizontal-divider-bar"></span>
@@ -41,26 +29,27 @@
 
         <div class="create-account-options">
             <span style="color: var(--color-subtext)">
-                {{ t('alreadyHaveAccountText') }}
+                {{ t('loginNewHereText') }}
             </span>
 
             <router-link
-                to="/auth/login"
+                to="/auth/register"
                 style="color: var(--vt-c-highlight); cursor: pointer; text-decoration: none"
             >
-                {{ t('registerReturnText') }}
+                {{ t('loginCreateAccountText') }}
             </router-link>
         </div>
     </div>
 </template>
 
 <style>
-    #register-page-content {
+    #login-page-content {
         display: flex;
         flex-direction: column;
         align-items: center;
         gap: 24px;
-        width: 80%;
+        width: 100%;
+        max-width: 300px;
         box-sizing: border-box;
     }
 </style>
